Drop valuesDirectoryURL entirely
valuesDirectoryURL was only needed when Argo did not support referencing
the values files inside the pattern's git repo root. This stopped
being the case around Argo 2.3 [1]. I.e. Now any value
file that starts with '/' is assumed to reference the git repo's root.

The main reason we do this, is because argo tends to cache values files
over http and ignore refresh requests (needs a hard-refresh), so the
expectation of doing gitops and tweaking values files which are then
fetched over http, breaks down somewhat. [2]

We have three possible options to fix this:
1. We ignore the issue [2] entirely. We keep things as is and we revert
   the operator change [3]. I.e. we always fetch the values files from
   URL (this keeps things as is and we unbreak IE when deployed via the
   operator), but we are more susceptible to the gitops values caching
   issue. We could potentially reevaluate this once [4] gets
   implemented.

2. We completely drop valuesDirectoryURL and we always reference local files.
   External charts need to override all the things. This is quite simple, but
   has a bit of an unknown as we do not know how painful this could be with
   certain external charts, as those get extracted in a separate folder and
   there is no access to the git-managed values-files and also because
   overrides won't allow for the use of variables in helm [5]

3. We add a bunch of ifs everywhere and use local files when: A) we're not an
   external chart and B) when valuesDirectoryURL is not defined

For the time being we choose option 2. and completely drop any mention
of valuesDirectoryURL. If this becomes a serious issue with external
charts, we can re-evaluate our option at that time.

[1] https://github.com/argoproj/argo-cd/blob/bb77664b6f0299bb332843bcd2524820c7ba1558/reposerver/repository/repository.go#L674
[2] https://issues.redhat.com/browse/GITOPS-2069
[3] https://github.com/hybrid-cloud-patterns/patterns-operator/commit/ba3c35a15dd5f239c7a7d337747b01d494278429
[4] https://github.com/argoproj/argo-cd/pull/8322/files
[5] https://github.com/helm/helm/issues/2492