Add initial multi-source support
This change adds initial multiSource support to patterns' applications.
The way this works is that nothing changes for applications defined in
values-*.yaml by default. So all patterns will work as usual.

What can change with this patch is that applications can slowly migrate
towards using multi source by changing an app definition from:

    acm:
      name: acm
      path: common/acm

To:

    acm:
      name: acm
      chart: acm
      chartVersion: 0.0.*

So any time we have a `chart` field with a `chartVersion` and no
`repoURL` defined, the clustergroup chart will create a multisource
application with the values files taken from the patterns git repo
and the helm chart from https://charts.validatedpatterns.io/ using
the `chartVersion` defined in the application.

For example the above acm app would prodice the following:

    project: hub
    sources:
    - ref: patternref
      repoURL: https://github.com/mbaldessari/multicloud-gitops
      targetRevision: multisource-test2
    - chart: acm
      helm:
        ignoreMissingValueFiles: true
        parameters:
        - name: global.repoURL
          value: $ARGOCD_APP_SOURCE_REPO_URL
        - name: global.targetRevision
          value: $ARGOCD_APP_SOURCE_TARGET_REVISION
        - name: global.namespace
          value: $ARGOCD_APP_NAMESPACE
        - name: global.pattern
          value: pattern-sample
        - name: global.clusterDomain
          value: mcg-hub.blueprints.rhecoeng.com
        - name: global.clusterVersion
          value: "4.13"
        - name: global.clusterPlatform
          value: AWS
        - name: global.hubClusterDomain
          value: apps.mcg-hub.blueprints.rhecoeng.com
        - name: global.localClusterDomain
          value: apps.mcg-hub.blueprints.rhecoeng.com
        valueFiles:
        - $patternref/values-global.yaml
        - $patternref/values-hub.yaml
        - $patternref/values-AWS.yaml
        - $patternref/values-AWS-4.13.yaml
        - $patternref/values-AWS-hub.yaml
        - $patternref/values-4.13-hub.yaml
      repoURL: https://charts.validatedpatterns.io/
      targetRevision: 0.0.*

Note that this depends on the operator supporting multiSource
applications (version > 0.0.17).