Only create subscriptions if they have a name
Our CI pipelines use HELM_EXTRA_OPTS to override the value of pre-release
subscriptions. These extra values are baked into the pattern itself and
are ultimately propagated to the applications argocd maintains on our
edge clusters as well. So a value like `clusterGroup.subscriptions.acm.channel`
will lead to the clustergroup chart trying to deploy a subscription but without
a name since `clusterGroup.subscriptions.acm.name` IS NOT set in values-group-one.yaml
in the pattern's repo. Argo is then unable to create this Subscription in the edge
cluster since it is a malformed Kubernetes object without a name and then nothing
from the clustergroup is created in the edge cluster.