feat: add delete global value as top level variable for child charts
We want to pass a different value to child charts if we are on the
hub, since we want acm to first remove all the child applications
before we actually remove acm.
Change it to a string, and pass directly to child charts.
Acm chart will handle further logic, from clustergroup we only care if
we need to render the applications or not.