Make clusterGroup.isHubCluster optional
We can figure out if a node is the hub or not by omparing
global.localClusterDomain and global.hubClusterDomain. If those are the
same we're on the hub, if they are not we're not on the hub.

Keep supporting the clusterGroup.isHubCluster variable with precedence
in order to not break any existing pattern.

Tested on 4.19 and 4.20 with mcg hub/spoke